move font rendering into RenderBuilder and fix dashed lines.

+++ src/main/kotlin/com/lambda/interaction/managers/breaking/BreakConfig.kt
@@ -73,7 +73,7 @@ interface BreakConfig : ActionConfig, ISettingGroup {
     val renders: Boolean
     val fill: Boolean
     val outline: Boolean
-    val outlineWidth: Int
+    val outlineWidth: Float
     val animation: AnimationMode
 
     val dynamicFillColor: Boolean
